Check for reserved names in lun for WRITE
In hompack fixpdf.f there's the line

```
  write(trace, 117) ...
```

f2cl was translating this to

```
(f2cl-lib:fformat trace ...)
```

But f2cl already renamed the variable `trace` to `trace$` so
`f2cl-lib:format` was referring to an unknown variable.

So when we parse the lun we need to check for reserved lisp names so
we get the right name.