Fix error in determining if PARAMETER is used
When checking to see if the value of a `PARAMETER` used other
`PARAMETER` variables, we didn't handle the case when the value was an
expression.  In this case we would sometimes not see the `PARAMETER`
in use and erroneously think it's not used and then not initialize it.

So, to make things easy, just flatten the expression and look through
the list to see if it contains any use of a `PARAMETER`.

Issue found when compiling toms/698/dcuhre.f