Get changes from Maxima
Maxima has made some changes to src/numerical/f2cl-lib.lisp.  This is
the same as our macros.l, so incorporate as many of these changes as
possible.

The only change I left out was in formatting where Maxima changed
`numberp` to `bigfloat:numberp`.

Have to think about that one.