Clean up minpack.system and define system to create lisp files
Get rid of the RCS log comments; they just clutter up the code.

Get rid of the logical pathname and define `*minpack-root*` as the
root of the minpack files.

Add a defsystem that just translates the Fortran files to Lisp,
putting the result in minpack/lib.

Tested this by building minpack-tests and running it.  It works.