Update quadpack
Use `*load-truename*` instead of `*load-pathname*`.

For some reason we weren't compiling a bunch of files that xerror
calls such as j4save, xerprt, fdump, and so on.  Added these files and
added them as a dependency of xerror.

Don't remember why we set the package to slatec for some files.
Remove that.

quadpack-tests runs fine.