Fix errors and update values
We accidentally deleted the NEGEP value so restore it.  We also had
the wrong sign for MACHEP.  The value if IRND should be 5 but we were
computing the value to be 2, which is incorrect.

Update the value of EPS and EPSNEG to match more closely what Fortran
machar returns.

For the record, I compiled a simple Fortran main program to call
Fortran MACHAR and printed the results.  Here is the output:

           2
          53
           5
           0
         -52
         -53
          11
       -1022
        1024
   2.2204460492503131E-016
   1.1102230246251565E-016
   2.2250738585072014E-308
   1.7976931348623157E+308

This lisp version currently returns

2
53
5
0
-52
-53
11
-1022
1024
2.2204460492503136d-16
1.1102230246251568d-16
2.2250738585072014d-308
1.7976931348623157d308

So we match everything except EPS and EPSNEG.