Fix #16:  Add ignore declaration for other unused variables
To ignore other unused variables, we need to see if they're used in
the body.  This is complicated by the fact that sometimes `fref` uses
the variable as an upper bound for an index.  But strangely, the
bounds check is done when compiling the lisp code after f2cl-lib.lisp
has been loaded.